fix test http server and clean up unit tests
