+++ tests/conftest.py
@@ -1,25 +1,17 @@
 import contextlib
-import json
-import os
-import socket
 import threading
-from http.server import BaseHTTPRequestHandler, HTTPServer
-from time import sleep
+from http.server import HTTPServer
 from typing import Optional
-from unittest.mock import patch
 
-import boto3
 import pytest
+import requests
 from alembic.config import main
 from docker.models.containers import ContainerCollection
 from fastapi.testclient import TestClient
-from sqlalchemy import create_engine
-from sqlalchemy.orm import Session, sessionmaker
 
 from app.routes import is_admin, is_service_account
 from app.settings.globals import (
     AURORA_JOB_QUEUE,
-    AWS_REGION,
     DATA_LAKE_JOB_QUEUE,
     GDAL_PYTHON_JOB_DEFINITION,
     PIXETL_JOB_DEFINITION,
@@ -27,128 +19,19 @@ from app.settings.globals import (
     POSTGRESQL_CLIENT_JOB_DEFINITION,
     TILE_CACHE_JOB_DEFINITION,
     TILE_CACHE_JOB_QUEUE,
-    WRITER_DBNAME,
-    WRITER_HOST,
-    WRITER_PASSWORD,
-    WRITER_PORT,
-    WRITER_USERNAME,
 )
 
-from moto import mock_batch, mock_iam, mock_ecs, mock_ec2, mock_logs  # isort:skip
+from . import (
+    AWSMock,
+    MemoryServer,
+    is_admin_mocked,
+    is_service_account_mocked,
+    session,
+    setup_clients,
+)
 
-SessionLocal: Optional[Session] = None
-LOG_GROUP = "/aws/batch/job"
-ROOT = os.environ["ROOT"]
-
-
-class AWSMock(object):
-    mocks = {
-        "batch": mock_batch,
-        "iam": mock_iam,
-        "ecs": mock_ecs,
-        "ec2": mock_ec2,
-        "logs": mock_logs,
-    }
-
-    def __init__(self, *services):
-        self.mocked_services = dict()
-        for service in services:
-            mocked_service = self.mocks[service]()
-            mocked_service.start()
-            client = boto3.client(service, region_name=AWS_REGION)
-            self.mocked_services[service] = {
-                "client": client,
-                "mock": mocked_service,
-            }
-        self.add_log_group(LOG_GROUP)
-
-    def stop_services(self):
-        for service in self.mocked_services.keys():
-            self.mocked_services[service]["mock"].stop()
-
-    def add_log_group(self, log_group_name):
-        self.mocked_services["logs"]["client"].create_log_group(
-            logGroupName=log_group_name
-        )
-
-    def add_compute_environment(self, compute_name, subnet_id, sg_id, iam_arn):
-        return self.mocked_services["batch"]["client"].create_compute_environment(
-            computeEnvironmentName=compute_name,
-            type="MANAGED",
-            state="ENABLED",
-            computeResources={
-                "type": "EC2",
-                "minvCpus": 0,
-                "maxvCpus": 2,
-                "desiredvCpus": 2,
-                "instanceTypes": ["t2.small", "t2.medium"],
-                "imageId": "some_image_id",
-                "subnets": [subnet_id],
-                "securityGroupIds": [sg_id],
-                "ec2KeyPair": "string",
-                "instanceRole": iam_arn.replace("role", "instance-profile"),
-                "tags": {"string": "string"},
-                "bidPercentage": 100,
-                "spotIamFleetRole": "string",
-            },
-            serviceRole=iam_arn,
-        )
-
-    def add_job_queue(self, job_queue_name, env_arn):
-        return self.mocked_services["batch"]["client"].create_job_queue(
-            jobQueueName=job_queue_name,
-            state="ENABLED",
-            priority=123,
-            computeEnvironmentOrder=[{"order": 123, "computeEnvironment": env_arn}],
-        )
-
-    def add_job_definition(self, job_definition_name, docker_image):
-
-        return self.mocked_services["batch"]["client"].register_job_definition(
-            jobDefinitionName=job_definition_name,
-            type="container",
-            containerProperties={
-                "image": f"{docker_image}:latest",
-                "vcpus": 1,
-                "memory": 128,
-                "environment": [
-                    {"name": "AWS_ACCESS_KEY_ID", "value": "testing"},
-                    {"name": "AWS_SECRET_ACCESS_KEY", "value": "testing"},
-                    {"name": "DEBUG", "value": "1"},
-                ],
-                "volumes": [
-                    {
-                        "host": {"sourcePath": f"{ROOT}/tests/fixtures/aws"},
-                        "name": "aws",
-                    }
-                ],
-                "mountPoints": [
-                    {
-                        "sourceVolume": "aws",
-                        "containerPath": "/root/.aws",
-                        "readOnly": True,
-                    }
-                ],
-            },
-        )
-
-    def print_logs(self):
-        resp = self.mocked_services["logs"]["client"].describe_log_streams(
-            logGroupName=LOG_GROUP
-        )
-
-        for stream in resp["logStreams"]:
-            ls_name = stream["logStreamName"]
-
-            stream_resp = self.mocked_services["logs"]["client"].get_log_events(
-                logGroupName=LOG_GROUP, logStreamName=ls_name
-            )
-
-            print(f"-------- LOGS FROM {ls_name} --------")
-            for event in stream_resp["events"]:
-                print(event["message"])
-
-
+# We overwrite endpoint_url directly in the app.
+# Keeping this around for now, just in case we want to revert back to fixtures.
 # @pytest.fixture(autouse=True)
 # def moto_s3():
 #     with patch(
@@ -173,7 +56,7 @@ def batch_client():
 
     ContainerCollection.run = patch_run
 
-    vpc_id, subnet_id, sg_id, iam_arn = _setup(
+    vpc_id, subnet_id, sg_id, iam_arn = setup_clients(
         aws_mock.mocked_services["ec2"]["client"],
         aws_mock.mocked_services["iam"]["client"],
     )
@@ -206,53 +89,6 @@ def batch_client():
     aws_mock.stop_services()
 
 
-def _setup(ec2_client, iam_client):
-    """
-    Do prerequisite setup
-    :return: VPC ID, Subnet ID, Security group ID, IAM Role ARN
-    :rtype: tuple
-    """
-    resp = ec2_client.create_vpc(CidrBlock="172.30.0.0/24")
-    vpc_id = resp["Vpc"]["VpcId"]
-    resp = ec2_client.create_subnet(
-        AvailabilityZone="us-east-1a", CidrBlock="172.30.0.0/25", VpcId=vpc_id
-    )
-    subnet_id = resp["Subnet"]["SubnetId"]
-    resp = ec2_client.create_security_group(
-        Description="test_sg_desc", GroupName="test_sg", VpcId=vpc_id
-    )
-    sg_id = resp["GroupId"]
-
-    resp = iam_client.create_role(
-        RoleName="TestRole", AssumeRolePolicyDocument="some_policy"
-    )
-    iam_arn = resp["Role"]["Arn"]
-    iam_client.create_instance_profile(InstanceProfileName="TestRole")
-    iam_client.add_role_to_instance_profile(
-        InstanceProfileName="TestRole", RoleName="TestRole"
-    )
-
-    return vpc_id, subnet_id, sg_id, iam_arn
-
-
-@contextlib.contextmanager
-def session():
-    global SessionLocal
-
-    if SessionLocal is None:
-        db_conn = f"postgresql://{WRITER_USERNAME}:{WRITER_PASSWORD}@{WRITER_HOST}:{WRITER_PORT}/{WRITER_DBNAME}"  # pragma: allowlist secret
-        engine = create_engine(db_conn, pool_size=1, max_overflow=0)
-        SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-
-    db: Optional[Session] = None
-    try:
-        db = SessionLocal()
-        yield db
-    finally:
-        if db is not None:
-            db.close()
-
-
 @pytest.fixture(scope="session", autouse=True)
 def db():
     """
@@ -262,14 +98,6 @@ def db():
     """
     with contextlib.ExitStack() as stack:
         yield stack.enter_context(session())
-
-
-async def is_admin_mocked():
-    return True
-
-
-async def is_service_account_mocked():
-    return True
 
 
 @pytest.fixture(autouse=True)
@@ -291,70 +119,13 @@ def client():
     main(["--raiseerr", "downgrade", "base"])
 
 
-class MemoryServer(BaseHTTPRequestHandler):
-    requests_thus_far: list = []
-
-    def do_GET(self):
-        self.send_response(200)
-        self.send_header("Content-type", "application/json")
-        self.end_headers()
-        self.wfile.write(
-            json.dumps({"requests": self.requests_thus_far}).encode("utf-8")
-        )
-
-    def do_DELETE(self):
-        self.send_response(200)
-        self.send_header("Content-type", "application/json")
-        self.end_headers()
-        self.wfile.write(json.dumps({"foo": "bar"}).encode("utf-8"))
-        self.requests_thus_far = []
-
-    def do_PUT(self):
-        self.send_response(200)
-        self.send_header("Content-type", "application/json")
-        self.end_headers()
-        self.wfile.write(json.dumps({"foo": "bar"}).encode("utf-8"))
-
-        content_length = int(self.headers["Content-Length"])
-        put_data = self.rfile.read(content_length)
-
-        self.requests_thus_far.append(
-            {"path": str(self.path), "body": json.loads(str(put_data.decode("utf-8")))}
-        )
-
-    def do_PATCH(self):
-        self.send_response(200)
-        self.send_header("Content-type", "application/json")
-        self.end_headers()
-        self.wfile.write(json.dumps({"foo": "bar"}).encode("utf-8"))
-
-        content_length = int(self.headers["Content-Length"])
-        put_data = self.rfile.read(content_length)
-
-        self.requests_thus_far.append(
-            {"path": str(self.path), "body": json.loads(str(put_data.decode("utf-8")))}
-        )
-
-
-@pytest.fixture(scope="function")
+@pytest.fixture(scope="session")
 def httpd():
     server_class = HTTPServer
     handler_class = MemoryServer
     port = 9000
 
     httpd = server_class(("0.0.0.0", port), handler_class)
-
-    # httpd.allow_reuse_address = True
-    # httpd.socket.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-
-    # while port <= 9100:
-    #     try:
-    #         httpd = server_class(("0.0.0.0", port), handler_class)
-    #         break
-    #         # httpd.allow_reuse_address = True
-    #     except OSError:
-    #         # Port hasn't been let go yet
-    #         port += 1
 
     t = threading.Thread(target=httpd.serve_forever)  # , daemon=True)
     t.start()
@@ -363,3 +134,11 @@ def httpd():
 
     httpd.shutdown()
     t.join()
+
+
+@pytest.fixture(autouse=True)
+def flush_request_list(httpd):
+    """
+    Delete request cache before every test
+    """
+    requests.delete(f"http://localhost:{httpd.server_port}")
